requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.2,<25.0
drf-spectacular>=0.27,<1.0
python-dotenv>=1.0
pytest>=8.0
pytest-django>=4.8